Refactor experimental layout user menu
