+++ src/apps/experimental/components/AppToolbar/index.tsx
@@ -1,9 +1,7 @@
 import MenuIcon from '@mui/icons-material/Menu';
 import SearchIcon from '@mui/icons-material/Search';
-import Avatar from '@mui/material/Avatar';
 import Box from '@mui/material/Box';
 import IconButton from '@mui/material/IconButton';
-import { useTheme } from '@mui/material/styles';
 import Toolbar from '@mui/material/Toolbar';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
@@ -14,23 +12,20 @@ import appIcon from 'assets/img/icon-transparent.png';
 import { useApi } from 'hooks/useApi';
 import globalize from 'scripts/globalize';
 
-import { ID as UserMenuId } from './AppUserMenu';
-import AppTabs from './tabs/AppTabs';
-import { isDrawerPath } from './drawers/AppDrawer';
+import AppTabs from '../tabs/AppTabs';
+import { isDrawerPath } from '../drawers/AppDrawer';
+import UserMenuButton from './UserMenuButton';
 
 interface AppToolbarProps {
     isDrawerOpen: boolean
     onDrawerButtonClick: (event: React.MouseEvent<HTMLElement>) => void
-    onUserButtonClick: (event: React.MouseEvent<HTMLElement>) => void
 }
 
 const AppToolbar: FC<AppToolbarProps> = ({
     isDrawerOpen,
-    onDrawerButtonClick,
-    onUserButtonClick
+    onDrawerButtonClick
 }) => {
-    const theme = useTheme();
-    const { api, user } = useApi();
+    const { user } = useApi();
     const isUserLoggedIn = Boolean(user);
     const location = useLocation();
 
@@ -108,31 +103,7 @@ const AppToolbar: FC<AppToolbarProps> = ({
                     </Box>
 
                     <Box sx={{ flexGrow: 0 }}>
-                        <Tooltip title={globalize.translate('UserMenu')}>
-                            <IconButton
-                                size='large'
-                                edge='end'
-                                aria-label={globalize.translate('UserMenu')}
-                                aria-controls={UserMenuId}
-                                aria-haspopup='true'
-                                onClick={onUserButtonClick}
-                                color='inherit'
-                                sx={{ padding: 0 }}
-                            >
-                                <Avatar
-                                    alt={user?.Name || undefined}
-                                    src={
-                                        api && user?.Id ?
-                                            `${api.basePath}/Users/${user.Id}/Images/Primary?tag=${user.PrimaryImageTag}` :
-                                            undefined
-                                    }
-                                    sx={{
-                                        bgcolor: theme.palette.primary.dark,
-                                        color: 'inherit'
-                                    }}
-                                />
-                            </IconButton>
-                        </Tooltip>
+                        <UserMenuButton />
                     </Box>
                 </>
             )}
